+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Upgrading to 3.6.x versions requires that you upgrade to latest 3.5.x version first.
 
+- Remove last PEAR dependency: `Math_Stats` (@glensc, #509)
+
 [3.6.4]: https://github.com/eventum/eventum/compare/v3.6.3...master
 
 ## [3.6.3] - 2019-03-21